refactor(routing): establish api/rest/ package structure
Move RouteRegistry and REST-specific types into api/rest/ package:
- api/rest/routing.py — RouteRegistry (moved from api/routing.py)
- api/rest/types.py — canonical location for REST types
- api/rest/middleware/ — placeholder for middleware extraction
- api/types.py — re-exports REST types for backward compatibility

This creates a symmetric structure with api/gql/ and enables
separate CLAUDE.md guidelines per API protocol.
