+++ src/ai/backend/manager/api/types.py
@@ -1,34 +1,31 @@
 from __future__ import annotations
 
-from collections.abc import Awaitable, Callable, Iterable, Mapping
+from collections.abc import Callable
 from contextlib import AbstractAsyncContextManager
 from typing import (
     TYPE_CHECKING,
 )
 
-import aiohttp_cors
-from aiohttp import web
-from aiohttp.typedefs import Middleware
+# REST types — canonical location is api.rest.types.
+# Re-exported here for backward compatibility during migration.
+from .rest.types import (
+    AppCreator as AppCreator,
+)
+from .rest.types import (
+    CORSOptions as CORSOptions,
+)
+from .rest.types import (
+    RouteMiddleware as RouteMiddleware,
+)
+from .rest.types import (
+    WebMiddleware as WebMiddleware,
+)
+from .rest.types import (
+    WebRequestHandler as WebRequestHandler,
+)
 
 if TYPE_CHECKING:
     from .context import RootContext
 
-
-type WebRequestHandler = Callable[
-    [web.Request],
-    Awaitable[web.StreamResponse],
-]
-type WebMiddleware = Middleware
-
-type CORSOptions = Mapping[str, aiohttp_cors.ResourceOptions]
-type AppCreator = Callable[
-    [CORSOptions],
-    tuple[web.Application, Iterable[WebMiddleware]],
-]
-
-type RouteMiddleware = Callable[
-    [WebRequestHandler],
-    WebRequestHandler,
-]
 
 type CleanupContext = Callable[["RootContext"], AbstractAsyncContextManager[None]]
